+++ changelog.md
@@ -3,8 +3,9 @@
 ## Upcoming
 * Iron armor Toughness increased (~~2~~ -> 4 total)
 * Weighted Equipment (renamed from Weighted Armor) 
-  * Shield Slowdown moved from Stats to Weighted Equipment 
+  * Shield Slowdown moved from Stats to Weighted Equipment and the slowdown is now configurable
   * Shield slowdown reduced (25% -> 18% slowdown)
+  * Reduced armor slowdown 
 * Taking damage from non-entities (e.g. falling, poison) will no longer stop eating/drinking
 
 ## 2.9.6